some bug fixed

+++ src/app/controllers/category.controllers.js
@@ -1,7 +1,7 @@
 const httpStatus = require("http-status");
 const sendResponse = require("../../shared/send.response");
 const ApiError = require("../../errors/ApiError");
-const { getAllCategoryService, checkACategoryExits, postCategoryServices, updateCategoryServices, deleteCategoryServices } = require("../services/category.services");
+const { getAllCategoryService, checkACategoryExits, postCategoryServices, updateCategoryServices, deleteCategoryServices, checkACategoryExitsInSubCategory, checkACategoryExitsInProducts, checkACategoryExitsInCategoryWhenUpdate } = require("../services/category.services");
 
 // get all Category
 exports.getAllCategory = async (req, res, next) => {
@@ -48,6 +48,10 @@ exports.postCategory = async (req, res, next) => {
 exports.updateCategoryInfo = async (req, res, next) => {
     try {
         const data = req.body;
+        const checkExist = await checkACategoryExitsInCategoryWhenUpdate(data?.category);
+        if(checkExist){
+            throw new ApiError(400, 'Previously Added !')
+        }
         const result= await updateCategoryServices(data);
         if (result?.modifiedCount > 0) {
             sendResponse(res, {
@@ -68,6 +72,14 @@ exports.updateCategoryInfo = async (req, res, next) => {
 exports.deleteACategoryInfo = async (req, res, next) => {
     try {
         const id = req.body._id;
+        const existSubCategory = await checkACategoryExitsInSubCategory(id);
+        if(existSubCategory?.length > 0){
+            throw new ApiError(400, 'This category is exist in sub category !');
+        }
+        const existProduct = await checkACategoryExitsInProducts(id);
+        if(existProduct?.length > 0){
+            throw new ApiError(400, 'This category is exist in products !');
+        }
         const result = await deleteCategoryServices(id);
         if (result?.deletedCount > 0) {
             sendResponse(res, {
